Fixed test cases to cast to short

+++ src/test/java/CatapultTest.java
@@ -37,17 +37,17 @@ public class CatapultTest extends ApplicationTest {
     public void selectMonster() {
         Arena arena = new Arena(new Label(), new AnchorPane());
         Image img = new Image("/unicorn.png");
-        Coordinates destination = new Coordinates(470,10);
+        Coordinates destination = new Coordinates((short) 470, (short) 10);
         PriorityQueue<Monster> testList = new PriorityQueue<>();
         LinkedList<ExistsInArena> selectList = new LinkedList<>();
         LinkedList<Monster> answerList = new LinkedList<>();
         Coordinates target;
 
         //test shooting range valid
-        Catapult catapult = new Catapult(arena, new Coordinates(370,10));
-        Monster m1 = new Unicorn(arena, new Coordinates(110, 110), destination, new ImageView(img), 1);testList.add(m1);arena.addObject(m1);
-        Monster m2 = new Unicorn(arena, new Coordinates(130, 110), destination, new ImageView(img), 1);testList.add(m2);arena.addObject(m2);
-        Monster m3 = new Unicorn(arena, new Coordinates(450, 10), destination, new ImageView(img), 1);testList.add(m3);arena.addObject(m3);
+        Catapult catapult = new Catapult(arena, new Coordinates((short) 370,(short) 10));
+        Monster m1 = new Unicorn(arena, new Coordinates((short) 110, (short) 110), destination, new ImageView(img), 1);testList.add(m1);arena.addObject(m1);
+        Monster m2 = new Unicorn(arena, new Coordinates((short) 130, (short) 110), destination, new ImageView(img), 1);testList.add(m2);arena.addObject(m2);
+        Monster m3 = new Unicorn(arena, new Coordinates((short) 450, (short) 10), destination, new ImageView(img), 1);testList.add(m3);arena.addObject(m3);
         target=catapult.selectMonster(testList,selectList);
         answerList.add(m3);
         Assert.assertEquals(selectList,answerList);
@@ -57,8 +57,8 @@ public class CatapultTest extends ApplicationTest {
         selectList.clear();
 
         //test shoot monster with nearest but not most monsters
-        Monster m4 = new Unicorn(arena, new Coordinates(330, 10), destination, new ImageView(img), 1);testList.add(m4);arena.addObject(m4);
-        Monster m5 = new Unicorn(arena, new Coordinates(330, 10), destination, new ImageView(img), 1);testList.add(m5);arena.addObject(m5);
+        Monster m4 = new Unicorn(arena, new Coordinates((short) 330, (short) 10), destination, new ImageView(img), 1);testList.add(m4);arena.addObject(m4);
+        Monster m5 = new Unicorn(arena, new Coordinates((short) 330, (short) 10), destination, new ImageView(img), 1);testList.add(m5);arena.addObject(m5);
         target=catapult.selectMonster(testList,selectList);
         Assert.assertEquals(selectList,answerList);
         for (ExistsInArena m :selectList) {
@@ -67,8 +67,8 @@ public class CatapultTest extends ApplicationTest {
         selectList.clear();
 
         //test shoot monster with same nearest but most monster
-        Monster m6 = new Unicorn(arena, new Coordinates(430, 10), destination, new ImageView(img), 1);testList.add(m6);arena.addObject(m6);
-        Monster m7 = new Unicorn(arena, new Coordinates(470, 50), destination, new ImageView(img), 1);testList.add(m7);arena.addObject(m7);
+        Monster m6 = new Unicorn(arena, new Coordinates((short) 430, (short) 10), destination, new ImageView(img), 1);testList.add(m6);arena.addObject(m6);
+        Monster m7 = new Unicorn(arena, new Coordinates((short) 470, (short) 50), destination, new ImageView(img), 1);testList.add(m7);arena.addObject(m7);
         target=catapult.selectMonster(testList,selectList);
         answerList.add(m6);
         Assert.assertEquals(selectList,answerList);
